add split clients into different classes
