+++ javascript/modules/client/package.json
@@ -18,7 +18,7 @@
   "scripts": {
     "start": "tsdx watch",
     "build": "tsdx build",
-    "test": "TESTING=true tsdx test --verbose",
+    "test": "TESTING=true tsdx test --verbose --runInBand",
     "lint": "tsdx lint",
     "prepare": "tsdx build",
     "size": "size-limit",
