+++ javascript/modules/client/src/modules-addressList/estimation.ts
@@ -1,0 +1,81 @@
+import { Random } from "@aragon/sdk-common";
+import { ContextPlugin } from "../context-plugin";
+import { ClientCore } from "../internal/core";
+import { GasFeeEstimation } from "../internal/interfaces/common";
+import {
+  IClientAddressListEstimation,
+  ICreateProposalParams,
+  IExecuteProposalParams,
+  IVoteProposalParams,
+} from "../internal/interfaces/plugins";
+
+export class IClientAddressListEstimationModule extends ClientCore
+  implements IClientAddressListEstimation {
+  constructor(context: ContextPlugin) {
+    super(context);
+  }
+  
+  /**
+   * Estimates the gas fee of creating a proposal on the plugin
+   *
+   * @param {ICreateProposalParams} _params
+   * @return {*}  {Promise<GasFeeEstimation>}
+   * @memberof IClientAddressListEstimationModule
+   */
+  public createProposal(
+    _params: ICreateProposalParams,
+  ): Promise<GasFeeEstimation> {
+    const signer = this.web3.getConnectedSigner();
+    if (!signer) {
+      throw new Error("A signer is needed");
+    } else if (!signer.provider) {
+      throw new Error("A web3 provider is needed");
+    }
+    // TODO: Implement
+    return Promise.resolve(
+      this.web3.getApproximateGasFee(Random.getBigInt(BigInt(1500))),
+    );
+  }
+
+  /**
+   * Estimates the gas fee of casting a vote on a proposal
+   *
+   * @param {IVoteProposalParams} _params
+   * @return {*}  {Promise<GasFeeEstimation>}
+   * @memberof IClientAddressListEstimationModule
+   */
+  public voteProposal(_params: IVoteProposalParams): Promise<GasFeeEstimation> {
+    const signer = this.web3.getConnectedSigner();
+    if (!signer) {
+      throw new Error("A signer is needed");
+    } else if (!signer.provider) {
+      throw new Error("A web3 provider is needed");
+    }
+    // TODO: Implement
+    return Promise.resolve(
+      this.web3.getApproximateGasFee(Random.getBigInt(BigInt(1500))),
+    );
+  }
+
+  /**
+   * Estimates the gas fee of executing an AddressList proposal
+   *
+   * @param {IExecuteProposalParams} _params
+   * @return {*}  {Promise<GasFeeEstimation>}
+   * @memberof IClientAddressListEstimationModule
+   */
+  public executeProposal(
+    _params: IExecuteProposalParams,
+  ): Promise<GasFeeEstimation> {
+    const signer = this.web3.getConnectedSigner();
+    if (!signer) {
+      throw new Error("A signer is needed");
+    } else if (!signer.provider) {
+      throw new Error("A web3 provider is needed");
+    }
+    // TODO: Implement
+    return Promise.resolve(
+      this.web3.getApproximateGasFee(Random.getBigInt(BigInt(1500))),
+    );
+  }
+}
