+++ javascript/modules/client/test/integration/client-addressList.test.ts
@@ -1,143 +1,15 @@
 // @ts-ignore
-declare const describe, it, beforeAll, afterAll, expect, test;
+declare const describe, it, expect;
 
 import { JsonRpcProvider } from "@ethersproject/providers";
 import { Wallet } from "@ethersproject/wallet";
-import {
-  Client,
-  ClientAddressList,
-  Context,
-  ContextParams,
-  ContextPlugin,
-  SortDirection,
-} from "../../src";
-import * as ganacheSetup from "../../../../helpers/ganache-setup";
-import * as deployContracts from "../../../../helpers/deployContracts";
+import { ClientAddressList, Context, ContextPlugin } from "../../src";
 import { Client as IpfsClient } from "@aragon/sdk-ipfs";
 import { GraphQLClient } from "graphql-request";
 
-import {
-  ExecuteProposalStep,
-  IAddressListPluginInstall,
-  ICanVoteParams,
-  ICreateProposalParams,
-  IExecuteProposalParams,
-  IPluginSettings,
-  IProposalQueryParams,
-  IVoteProposalParams,
-  ProposalCreationSteps,
-  ProposalSortBy,
-  ProposalStatus,
-  VoteProposalStep,
-  VoteValues,
-} from "../../src/internal/interfaces/plugins";
-import {
-  bytesToHex,
-  InvalidAddressError,
-  InvalidAddressOrEnsError,
-} from "@aragon/sdk-common";
-
-const IPFS_API_KEY = process.env.IPFS_API_KEY ||
-  Buffer.from(
-    "YjQ3N1JoRUNmOHM4c2RNN1hya0xCczJ3SGM0a0NNd3BiY0ZDNTVLdCAg==",
-    "base64",
-  )
-    .toString()
-    .trim();
-
-const web3endpoints = {
-  working: [
-    "https://mainnet.infura.io/v3/94d2e8caf1bc4c4884af830d96f927ca",
-    "https://cloudflare-eth.com/",
-  ],
-  failing: ["https://bad-url-gateway.io/"],
-};
-const ipfsEndpoints = {
-  working: [
-    {
-      url: "https://testing-ipfs-0.aragon.network/api/v0",
-      headers: {
-        "X-API-KEY": IPFS_API_KEY || "",
-      },
-    },
-  ],
-  failing: [
-    {
-      url: "https://bad-url-gateway.io/",
-    },
-  ],
-};
-const grapqhlEndpoints = {
-  working: [
-    {
-      url:
-        "https://api.thegraph.com/subgraphs/name/aragon/aragon-zaragoza-goerli",
-    },
-  ],
-  failing: [{ url: "https://bad-url-gateway.io/" }],
-};
-
-const TEST_WALLET =
-  "0xdf57089febbacf7ba0bc227dafbffa9fc08a93fdc68e1e42411a14efcf23656e";
-const TEST_PROPOSAL_ID = "0x0bcaeb2da9d8ce246ac0864f081b04d4380b6b9c_0x0";
-const TEST_DAO_ADDDRESS = "0xa40fd495c454d19c87c4d615100307f0283c017c";
-const TEST_PLUGIN_ADDRESS = "0x0bcaeb2da9d8ce246ac0864f081b04d4380b6b9c";
-const TEST_INVALID_ADDRESS = "0x1nv4l1d_4ddr355";
-const TEST_NON_EXISTING_ADDRESS = "0x1234567890123456789012345678901234567890";
-
-const contextParams: ContextParams = {
-  network: "mainnet",
-  signer: new Wallet(TEST_WALLET),
-  daoFactoryAddress: "0x0123456789012345678901234567890123456789",
-  web3Providers: web3endpoints.working,
-  ipfsNodes: ipfsEndpoints.working,
-  graphqlNodes: grapqhlEndpoints.working,
-};
-
-const contextParamsFailing: ContextParams = {
-  network: "mainnet",
-  signer: new Wallet(TEST_WALLET),
-  daoFactoryAddress: "0x0123456789012345678901234567890123456789",
-  web3Providers: web3endpoints.failing,
-  ipfsNodes: ipfsEndpoints.failing,
-  graphqlNodes: grapqhlEndpoints.failing,
-};
-
-const contextParamsLocalChain: ContextParams = {
-  network: 31337,
-  signer: new Wallet(TEST_WALLET),
-  daoFactoryAddress: "0xf8065dD2dAE72D4A8e74D8BB0c8252F3A9acE7f9",
-  web3Providers: ["http://localhost:8545"],
-  ipfsNodes: [
-    {
-      url: "http:localhost:5001",
-    },
-    {
-      url: "http:localhost:5002",
-    },
-    {
-      url: "http:localhost:5003",
-    },
-  ],
-  graphqlNodes: [
-    {
-      url:
-        "https://api.thegraph.com/subgraphs/name/aragon/aragon-zaragoza-goerli",
-    },
-  ],
-};
+import { contextParams, contextParamsFailing } from "./constants";
 
 describe("Client", () => {
-  beforeAll(async () => {
-    const server = await ganacheSetup.start();
-    const daoFactory = await deployContracts.deploy(server);
-    contextParamsLocalChain.daoFactoryAddress = daoFactory.address;
-  });
-
-  afterAll(async () => {
-    await ganacheSetup.stop();
-  });
-
   describe("Client instances", () => {
     it("Should create a working client", async () => {
       const ctx = new Context(contextParams);
@@ -183,719 +55,6 @@ describe("Client", () => {
       // GraqphQl
       const graphqlStatus = await client.graphql.isUp();
       expect(graphqlStatus).toEqual(false);
-    });
-  });
-  describe("Proposal Creation", () => {
-    it("Should estimate the gas fees for creating a new proposal", async () => {
-      const ctx = new Context(contextParamsLocalChain);
-      const ctxPlugin = ContextPlugin.fromContext(ctx);
-      const client = new ClientAddressList(ctxPlugin);
-
-      const proposalParams: ICreateProposalParams = {
-        pluginAddress: "0x1234567890123456789012345678901234567890",
-        metadata: {
-          title: "Best Proposal",
-          summary: "this is the sumnary",
-          description: "This is a very long description",
-          resources: [
-            {
-              name: "Website",
-              url: "https://the.website",
-            },
-          ],
-          media: {
-            header: "https://no.media/media.jpeg",
-            logo: "https://no.media/media.jpeg",
-          },
-        },
-        actions: [],
-        creatorVote: VoteValues.YES,
-        startDate: new Date(),
-        endDate: new Date(),
-        executeOnPass: true,
-      };
-
-      const estimation = await client.estimation.createProposal(proposalParams);
-
-      expect(typeof estimation).toEqual("object");
-      expect(typeof estimation.average).toEqual("bigint");
-      expect(typeof estimation.max).toEqual("bigint");
-      expect(estimation.max).toBeGreaterThan(BigInt(0));
-      expect(estimation.max).toBeGreaterThan(estimation.average);
-    });
-    it("Should create a new proposal locally", async () => {
-      const ctx = new Context(contextParamsLocalChain);
-      const ctxPlugin = ContextPlugin.fromContext(ctx);
-      const addressListClient = new ClientAddressList(ctxPlugin);
-      const client = new Client(ctx);
-
-      // generate actions
-      const action = await client.encoding.withdrawAction(
-        "0x1234567890123456789012345678901234567890",
-        {
-          recipientAddress: "0x1234567890123456789012345678901234567890",
-          amount: BigInt(1),
-          reference: "test",
-        },
-      );
-
-      const proposalParams: ICreateProposalParams = {
-        pluginAddress: "0x1234567890123456789012345678901234567890",
-        metadata: {
-          title: "Best Proposal",
-          summary: "this is the sumnary",
-          description: "This is a very long description",
-          resources: [
-            {
-              name: "Website",
-              url: "https://the.website",
-            },
-          ],
-          media: {
-            header: "https://no.media/media.jpeg",
-            logo: "https://no.media/media.jpeg",
-          },
-        },
-        actions: [action],
-        creatorVote: VoteValues.YES,
-        startDate: new Date(),
-        endDate: new Date(),
-        executeOnPass: true,
-      };
-
-      for await (
-        const step of addressListClient.methods.createProposal(
-          proposalParams,
-        )
-      ) {
-        switch (step.key) {
-          case ProposalCreationSteps.CREATING:
-            expect(typeof step.txHash).toBe("string");
-            expect(step.txHash).toMatch(/^0x[A-Fa-f0-9]{64}$/i);
-            break;
-          case ProposalCreationSteps.DONE:
-            expect(typeof step.proposalId).toBe("string");
-            expect(step.proposalId).toMatch(/^0x[A-Fa-f0-9]{64}$/i);
-            break;
-          default:
-            throw new Error(
-              "Unexpected proposal creation step: " +
-                Object.keys(step).join(", "),
-            );
-        }
-      }
-    });
-  });
-
-  describe("Vote on a proposal", () => {
-    it("Should estimate the gas fees for casting a vote", async () => {
-      const ctx = new Context(contextParamsLocalChain);
-      const ctxPlugin = ContextPlugin.fromContext(ctx);
-      const client = new ClientAddressList(ctxPlugin);
-
-      const voteParams: IVoteProposalParams = {
-        pluginAddress: "0x1234567890123456789012345678901234567890",
-        proposalId: "0x1234567890123456789012345678901234567890",
-        vote: VoteValues.YES,
-      };
-
-      const estimation = await client.estimation.voteProposal(voteParams);
-
-      expect(typeof estimation).toEqual("object");
-      expect(typeof estimation.average).toEqual("bigint");
-      expect(typeof estimation.max).toEqual("bigint");
-      expect(estimation.max).toBeGreaterThan(BigInt(0));
-      expect(estimation.max).toBeGreaterThan(estimation.average);
-    });
-
-    it("Should vote on a proposal locally", async () => {
-      const ctx = new Context(contextParamsLocalChain);
-      const ctxPlugin = ContextPlugin.fromContext(ctx);
-      const client = new ClientAddressList(ctxPlugin);
-
-      const voteParams: IVoteProposalParams = {
-        pluginAddress: "0x1234567890123456789012345678901234567890",
-        proposalId: "0x1234567890123456789012345678901234567890",
-        vote: VoteValues.YES,
-      };
-
-      for await (const step of client.methods.voteProposal(voteParams)) {
-        switch (step.key) {
-          case VoteProposalStep.VOTING:
-            expect(typeof step.txHash).toBe("string");
-            expect(step.txHash).toMatch(/^0x[A-Fa-f0-9]{64}$/i);
-            break;
-          case VoteProposalStep.DONE:
-            expect(typeof step.voteId).toBe("string");
-            expect(step.voteId).toMatch(/^0x[A-Fa-f0-9]{64}$/i);
-            break;
-          default:
-            throw new Error(
-              "Unexpected vote proposal step: " + Object.keys(step).join(", "),
-            );
-        }
-      }
-    });
-  });
-
-  describe("Execute proposal", () => {
-    it("Should estimate the gas fees for executing a proposal", async () => {
-      const ctx = new Context(contextParamsLocalChain);
-      const ctxPlugin = ContextPlugin.fromContext(ctx);
-      const client = new ClientAddressList(ctxPlugin);
-
-      const executeParams: IExecuteProposalParams = {
-        pluginAddress: "0x1234567890123456789012345678901234567890",
-        proposalId: "0x1234567890123456789012345678901234567890",
-      };
-      const estimation = await client.estimation.executeProposal(executeParams);
-
-      expect(typeof estimation).toEqual("object");
-      expect(typeof estimation.average).toEqual("bigint");
-      expect(typeof estimation.max).toEqual("bigint");
-      expect(estimation.max).toBeGreaterThan(BigInt(0));
-      expect(estimation.max).toBeGreaterThan(estimation.average);
-    });
-
-    it("Should execute a local proposal", async () => {
-      const ctx = new Context(contextParamsLocalChain);
-      const ctxPlugin = ContextPlugin.fromContext(ctx);
-      const client = new ClientAddressList(ctxPlugin);
-
-      const executeParams: IExecuteProposalParams = {
-        pluginAddress: "0x1234567890123456789012345678901234567890",
-        proposalId: "0x1234567890123456789012345678901234567890",
-      };
-      for await (const step of client.methods.executeProposal(executeParams)) {
-        switch (step.key) {
-          case ExecuteProposalStep.EXECUTING:
-            expect(typeof step.txHash).toBe("string");
-            expect(step.txHash).toMatch(/^0x[A-Fa-f0-9]{64}$/i);
-            break;
-          case ExecuteProposalStep.DONE:
-            break;
-          default:
-            throw new Error(
-              "Unexpected execute proposal step: " +
-                Object.keys(step).join(", "),
-            );
-        }
-      }
-    });
-  });
-
-  describe("Can vote", () => {
-    it("Should check if an user can vote in an Address List proposal", async () => {
-      const ctx = new Context(contextParamsLocalChain);
-      const ctxPlugin = ContextPlugin.fromContext(ctx);
-      const client = new ClientAddressList(ctxPlugin);
-
-      const params: ICanVoteParams = {
-        address: "0x1234567890123456789012345678901234567890",
-        proposalId: "0x1234567890123456789012345678901234567890_0x1",
-        pluginAddress: "0x1234567890123456789012345678901234567890",
-      };
-      const canVote = await client.methods.canVote(params);
-      expect(typeof canVote).toBe("boolean");
-    });
-  });
-
-  describe("Action generators", () => {
-    it("Should create an AddressList client and generate a install entry", async () => {
-      const withdrawParams: IAddressListPluginInstall = {
-        settings: {
-          minDuration: 7200, // seconds
-          minTurnout: 0.5,
-          minSupport: 0.5,
-        },
-        addresses: [
-          "0x1234567890123456789012345678901234567890",
-          "0x2345678901234567890123456789012345678901",
-          "0x3456789012345678901234567890123456789012",
-          "0x4567890123456789012345678901234567890134",
-        ],
-      };
-
-      const installPluginItemItem = ClientAddressList.encoding
-        .getPluginInstallItem(
-          withdrawParams,
-        );
-
-      expect(typeof installPluginItemItem).toBe("object");
-      // what does this should be
-      expect(installPluginItemItem.data).toBeInstanceOf(Uint8Array);
-    });
-
-    it("Should create an AddressList client and fail to generate a plugin config action with an invalid address", async () => {
-      const ctx = new Context(contextParamsLocalChain);
-      const ctxPlugin = ContextPlugin.fromContext(ctx);
-      const client = new ClientAddressList(ctxPlugin);
-
-      const pluginConfigParams: IPluginSettings = {
-        minDuration: 100000,
-        minTurnout: 0.25,
-        minSupport: 0.51,
-      };
-
-      const pluginAddress = "0xinvalid_address";
-      expect(() =>
-        client.encoding.updatePluginSettingsAction(
-          pluginAddress,
-          pluginConfigParams,
-        )
-      ).toThrow("Invalid plugin address");
-    });
-
-    it("Should create an AddressList client and generate a plugin config action action", async () => {
-      const ctx = new Context(contextParamsLocalChain);
-      const ctxPlugin = ContextPlugin.fromContext(ctx);
-      const client = new ClientAddressList(ctxPlugin);
-
-      const pluginConfigParams: IPluginSettings = {
-        minDuration: 100000,
-        minTurnout: 0.25,
-        minSupport: 0.51,
-      };
-
-      const pluginAddress = "0x1234567890123456789012345678901234567890";
-
-      const installPluginItemItem = client.encoding.updatePluginSettingsAction(
-        pluginAddress,
-        pluginConfigParams,
-      );
-
-      expect(typeof installPluginItemItem).toBe("object");
-      // what does this should be
-      expect(installPluginItemItem.data).toBeInstanceOf(Uint8Array);
-      expect(installPluginItemItem.to).toBe(pluginAddress);
-    });
-
-    it("Should encode a add members action with an invalid plugin address and fail", async () => {
-      const ctx = new Context(contextParamsLocalChain);
-      const ctxPlugin = ContextPlugin.fromContext(ctx);
-      const client = new ClientAddressList(ctxPlugin);
-
-      const members: string[] = [
-        "0x1357924680135792468013579246801357924680",
-        "0x2468013579246801357924680135792468013579",
-        "0x0987654321098765432109876543210987654321",
-      ];
-
-      const pluginAddress = "0xinvalid_address";
-      expect(() => client.encoding.addMembersAction(pluginAddress, members))
-        .toThrow(new InvalidAddressError());
-    });
-    it("Should encode a add members action with an invalid member address and fail", async () => {
-      const ctx = new Context(contextParamsLocalChain);
-      const ctxPlugin = ContextPlugin.fromContext(ctx);
-      const client = new ClientAddressList(ctxPlugin);
-
-      const members: string[] = [
-        "0xinvalid_address",
-        "0x2468013579246801357924680135792468013579",
-        "0x0987654321098765432109876543210987654321",
-      ];
-
-      const pluginAddress = "0x1234567890123456789012345678901234567890";
-      expect(() => client.encoding.addMembersAction(pluginAddress, members))
-        .toThrow(new InvalidAddressError());
-    });
-    it("Should encode a add members action", async () => {
-      const ctx = new Context(contextParamsLocalChain);
-      const ctxPlugin = ContextPlugin.fromContext(ctx);
-      const client = new ClientAddressList(ctxPlugin);
-
-      const members: string[] = [
-        "0x1357924680135792468013579246801357924680",
-        "0x2468013579246801357924680135792468013579",
-        "0x0987654321098765432109876543210987654321",
-      ];
-
-      const pluginAddress = "0x1234567890123456789012345678901234567890";
-      const action = client.encoding.addMembersAction(pluginAddress, members);
-
-      expect(typeof action).toBe("object");
-      expect(action.data).toBeInstanceOf(Uint8Array);
-      expect(bytesToHex(action.data, true)).toBe(
-        "0x6496d3fc00000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000003000000000000000000000000135792468013579246801357924680135792468000000000000000000000000024680135792468013579246801357924680135790000000000000000000000000987654321098765432109876543210987654321",
-      );
-      expect(action.to).toBe(pluginAddress);
-    });
-    it("Should encode a remove members action with an invalid plugin address and fail", async () => {
-      const ctx = new Context(contextParamsLocalChain);
-      const ctxPlugin = ContextPlugin.fromContext(ctx);
-      const client = new ClientAddressList(ctxPlugin);
-
-      const members: string[] = [
-        "0x1357924680135792468013579246801357924680",
-        "0x2468013579246801357924680135792468013579",
-        "0x0987654321098765432109876543210987654321",
-      ];
-
-      const pluginAddress = "0xinvalid_address";
-      expect(() => client.encoding.removeMembersAction(pluginAddress, members))
-        .toThrow(new InvalidAddressError());
-    });
-    it("Should encode a remove members action with an invalid member address and fail", async () => {
-      const ctx = new Context(contextParamsLocalChain);
-      const ctxPlugin = ContextPlugin.fromContext(ctx);
-      const client = new ClientAddressList(ctxPlugin);
-
-      const members: string[] = [
-        "0xinvalid_address",
-        "0x2468013579246801357924680135792468013579",
-        "0x0987654321098765432109876543210987654321",
-      ];
-
-      const pluginAddress = "0x1234567890123456789012345678901234567890";
-      expect(() => client.encoding.removeMembersAction(pluginAddress, members))
-        .toThrow(new InvalidAddressError());
-    });
-    it("Should encode a remove members action", async () => {
-      const ctx = new Context(contextParamsLocalChain);
-      const ctxPlugin = ContextPlugin.fromContext(ctx);
-      const client = new ClientAddressList(ctxPlugin);
-
-      const members: string[] = [
-        "0x1357924680135792468013579246801357924680",
-        "0x2468013579246801357924680135792468013579",
-        "0x0987654321098765432109876543210987654321",
-      ];
-
-      const pluginAddress = "0x1234567890123456789012345678901234567890";
-      const action = client.encoding.removeMembersAction(
-        pluginAddress,
-        members,
-      );
-
-      expect(typeof action).toBe("object");
-      expect(action.data).toBeInstanceOf(Uint8Array);
-      expect(bytesToHex(action.data, true)).toBe(
-        "0xeff42f2e00000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000003000000000000000000000000135792468013579246801357924680135792468000000000000000000000000024680135792468013579246801357924680135790000000000000000000000000987654321098765432109876543210987654321",
-      );
-      expect(action.to).toBe(pluginAddress);
-    });
-  });
-
-  describe("Action decoders", () => {
-    it("Should decode the plugin settings from an update plugin settings action", async () => {
-      const ctx = new Context(contextParamsLocalChain);
-      const ctxPlugin = ContextPlugin.fromContext(ctx);
-      const client = new ClientAddressList(ctxPlugin);
-      const params: IPluginSettings = {
-        minDuration: 7200,
-        minTurnout: 0.5,
-        minSupport: 0.5,
-      };
-
-      const pluginAddress = "0x1234567890123456789012345678901234567890";
-
-      const updatePluginSettingsAction = client.encoding
-        .updatePluginSettingsAction(
-          pluginAddress,
-          params,
-        );
-      const decodedParams: IPluginSettings = client.decoding
-        .updatePluginSettingsAction(
-          updatePluginSettingsAction.data,
-        );
-
-      expect(decodedParams.minDuration).toBe(params.minDuration);
-      expect(decodedParams.minSupport).toBe(params.minSupport);
-      expect(decodedParams.minTurnout).toBe(params.minTurnout);
-    });
-
-    it("Should try to decode a invalid action and with the update plugin settings decoder return an error", async () => {
-      const ctx = new Context(contextParamsLocalChain);
-      const ctxPlugin = ContextPlugin.fromContext(ctx);
-      const client = new ClientAddressList(ctxPlugin);
-      const data = new Uint8Array([11, 22, 22, 33, 33, 33]);
-
-      expect(() => client.decoding.updatePluginSettingsAction(data)).toThrow(
-        `no matching function (argument="sighash", value="0x0b161621", code=INVALID_ARGUMENT, version=abi/5.6.0)`,
-      );
-    });
-    it("Should decode a add members action", async () => {
-      const ctx = new Context(contextParamsLocalChain);
-      const ctxPlugin = ContextPlugin.fromContext(ctx);
-      const client = new ClientAddressList(ctxPlugin);
-
-      const members: string[] = [
-        "0x1357924680135792468013579246801357924680",
-        "0x2468013579246801357924680135792468013579",
-        "0x0987654321098765432109876543210987654321",
-      ];
-
-      const pluginAddress = "0x1234567890123456789012345678901234567890";
-      const action = client.encoding.addMembersAction(pluginAddress, members);
-      const decodedMembers = client.decoding.addMembersAction(action.data);
-      expect(Array.isArray(decodedMembers)).toBe(true);
-      for (let i = 0; i < decodedMembers.length; i++) {
-        expect(typeof decodedMembers[i]).toBe("string");
-        expect(decodedMembers[i]).toBe(members[i]);
-      }
-    });
-    it("Should decode a remove members action", async () => {
-      const ctx = new Context(contextParamsLocalChain);
-      const ctxPlugin = ContextPlugin.fromContext(ctx);
-      const client = new ClientAddressList(ctxPlugin);
-
-      const members: string[] = [
-        "0x1357924680135792468013579246801357924680",
-        "0x2468013579246801357924680135792468013579",
-        "0x0987654321098765432109876543210987654321",
-      ];
-
-      const pluginAddress = "0x1234567890123456789012345678901234567890";
-      const action = client.encoding.removeMembersAction(
-        pluginAddress,
-        members,
-      );
-      const decodedMembers = client.decoding.removeMembersAction(action.data);
-      expect(Array.isArray(decodedMembers)).toBe(true);
-      for (let i = 0; i < decodedMembers.length; i++) {
-        expect(typeof decodedMembers[i]).toBe("string");
-        expect(decodedMembers[i]).toBe(members[i]);
-      }
-    });
-
-    it("Should get the function for a given action data", async () => {
-      const ctx = new Context(contextParamsLocalChain);
-      const ctxPlugin = ContextPlugin.fromContext(ctx);
-      const client = new ClientAddressList(ctxPlugin);
-      const params: IPluginSettings = {
-        minDuration: 7200,
-        minTurnout: 0.5,
-        minSupport: 0.5,
-      };
-
-      const pluginAddress = "0x1234567890123456789012345678901234567890";
-
-      const updatePluginSettingsAction = client.encoding
-        .updatePluginSettingsAction(
-          pluginAddress,
-          params,
-        );
-      const iface = client.decoding.findInterface(
-        updatePluginSettingsAction.data,
-      );
-      expect(iface?.id).toBe("function changeVoteConfig(uint64,uint64,uint64)");
-      expect(iface?.functionName).toBe("changeVoteConfig");
-      expect(iface?.hash).toBe("0x634fe2fb");
-    });
-
-    it("Should try to get the function of an invalid data and return null", async () => {
-      const ctx = new Context(contextParamsLocalChain);
-      const ctxPlugin = ContextPlugin.fromContext(ctx);
-      const client = new ClientAddressList(ctxPlugin);
-      const data = new Uint8Array([11, 22, 22, 33, 33, 33]);
-      const iface = client.decoding.findInterface(data);
-      expect(iface).toBe(null);
-    });
-  });
-
-  describe("Data retrieval", () => {
-    it("Should get the list of members that can vote in a proposal", async () => {
-      const ctx = new Context(contextParams);
-      const ctxPlugin = ContextPlugin.fromContext(ctx);
-      const client = new ClientAddressList(ctxPlugin);
-
-      const wallets = await client.methods.getMembers(
-        TEST_DAO_ADDDRESS,
-      );
-
-      expect(Array.isArray(wallets)).toBe(true);
-      expect(wallets.length).toBeGreaterThan(0);
-      expect(typeof wallets[0]).toBe("string");
-      expect(wallets[0]).toMatch(/^0x[A-Fa-f0-9]{40}$/i);
-    });
-    it("Should fetch the given proposal", async () => {
-      const ctx = new Context(contextParams);
-      const ctxPlugin = ContextPlugin.fromContext(ctx);
-      const client = new ClientAddressList(ctxPlugin);
-
-      const proposalId = TEST_PROPOSAL_ID;
-      const proposal = await client.methods.getProposal(proposalId);
-
-      expect(typeof proposal).toBe("object");
-      expect(proposal === null).toBe(false);
-      if (proposal) {
-        expect(proposal.id).toBe(proposalId);
-        expect(typeof proposal.id).toBe("string");
-        expect(proposal.id).toMatch(/^0x[A-Fa-f0-9]{40}_0x[A-Fa-f0-9]{1,}$/i);
-        expect(typeof proposal.dao.address).toBe("string");
-        expect(proposal.dao.address).toMatch(/^0x[A-Fa-f0-9]{40}$/i);
-        expect(typeof proposal.dao.name).toBe("string");
-        expect(typeof proposal.creatorAddress).toBe("string");
-        expect(proposal.creatorAddress).toMatch(/^0x[A-Fa-f0-9]{40}$/i);
-        // check metadata
-        expect(typeof proposal.metadata.title).toBe("string");
-        expect(typeof proposal.metadata.summary).toBe("string");
-        expect(typeof proposal.metadata.description).toBe("string");
-        expect(Array.isArray(proposal.metadata.resources)).toBe(true);
-        for (let i = 0; i < proposal.metadata.resources.length; i++) {
-          const resource = proposal.metadata.resources[i];
-          expect(typeof resource.name).toBe("string");
-          expect(typeof resource.url).toBe("string");
-        }
-        if (proposal.metadata.media) {
-          if (proposal.metadata.media.header) {
-            expect(typeof proposal.metadata.media.header).toBe("string");
-          }
-          if (proposal.metadata.media.logo) {
-            expect(typeof proposal.metadata.media.logo).toBe("string");
-          }
-        }
-        expect(proposal.startDate instanceof Date).toBe(true);
-        expect(proposal.endDate instanceof Date).toBe(true);
-        expect(proposal.creationDate instanceof Date).toBe(true);
-        expect(Array.isArray(proposal.actions)).toBe(true);
-        // actions
-        for (let i = 0; i < proposal.actions.length; i++) {
-          const action = proposal.actions[i];
-          expect(action.data instanceof Uint8Array).toBe(true);
-          expect(typeof action.to).toBe("string");
-          expect(typeof action.value).toBe("bigint");
-        }
-        // result
-        expect(typeof proposal.result.yes).toBe("number");
-        expect(typeof proposal.result.no).toBe("number");
-        expect(typeof proposal.result.abstain).toBe("number");
-        // setttings
-        expect(typeof proposal.settings.duration).toBe("number");
-        expect(typeof proposal.settings.minSupport).toBe("number");
-        expect(typeof proposal.settings.minTurnout).toBe("number");
-        // TODO
-        // enable this tests when the subgrph have the correcto precision digits
-        // expect(
-        //   proposal.settings.minSupport >= 0 &&
-        //     proposal.settings.minSupport <= 1,
-        // ).toBe(true);
-        // expect(
-        //   proposal.settings.minTurnout >= 0 &&
-        //     proposal.settings.minTurnout <= 1,
-        // ).toBe(true);
-        // token
-        expect(typeof proposal.totalVotingWeight).toBe("number");
-        expect(Array.isArray(proposal.votes)).toBe(true);
-        for (let i = 0; i < proposal.votes.length; i++) {
-          const vote = proposal.votes[i];
-          expect(typeof vote.address).toBe("string");
-          expect(vote.address).toMatch(/^0x[A-Fa-f0-9]{40}$/i);
-          expect(typeof vote.vote).toBe("number");
-        }
-      }
-    });
-    it("Should fetch the given proposal and fail because the proposal does not exist", async () => {
-      const ctx = new Context(contextParams);
-      const ctxPlugin = ContextPlugin.fromContext(ctx);
-      const client = new ClientAddressList(ctxPlugin);
-
-      const proposalId = TEST_NON_EXISTING_ADDRESS + "_0x0";
-      const proposal = await client.methods.getProposal(proposalId);
-
-      expect(proposal === null).toBe(true);
-    });
-    it("Should get a list of proposals filtered by the given criteria", async () => {
-      const ctx = new Context(contextParams);
-      const ctxPlugin = ContextPlugin.fromContext(ctx);
-      const client = new ClientAddressList(ctxPlugin);
-      const limit = 5;
-      const status = ProposalStatus.EXECUTED;
-      const params: IProposalQueryParams = {
-        limit,
-        sortBy: ProposalSortBy.CREATED_AT,
-        direction: SortDirection.ASC,
-        status,
-      };
-      const proposals = await client.methods.getProposals(params);
-
-      expect(Array.isArray(proposals)).toBe(true);
-      expect(proposals.length <= limit).toBe(true);
-      for (let i = 0; i < proposals.length; i++) {
-        const proposal = proposals[i];
-        expect(typeof proposal.id).toBe("string");
-        expect(proposal.id).toMatch(/^0x[A-Fa-f0-9]{40}_0x[A-Fa-f0-9]{1,}$/i);
-        expect(typeof proposal.dao.address).toBe("string");
-        expect(proposal.dao.address).toMatch(/^0x[A-Fa-f0-9]{40}$/i);
-        expect(typeof proposal.dao.name).toBe("string");
-        expect(typeof proposal.creatorAddress).toBe("string");
-        expect(proposal.creatorAddress).toMatch(/^0x[A-Fa-f0-9]{40}$/i);
-        expect(typeof proposal.metadata.title).toBe("string");
-        expect(typeof proposal.metadata.summary).toBe("string");
-        expect(proposal.startDate instanceof Date).toBe(true);
-        expect(proposal.endDate instanceof Date).toBe(true);
-        expect(proposal.status).toBe(status);
-        // result
-        expect(typeof proposal.result.yes).toBe("number");
-        expect(typeof proposal.result.no).toBe("number");
-        expect(typeof proposal.result.abstain).toBe("number");
-      }
-    });
-    it("Should get a list of proposals from a specific dao", async () => {
-      const ctx = new Context(contextParams);
-      const ctxPlugin = ContextPlugin.fromContext(ctx);
-      const client = new ClientAddressList(ctxPlugin);
-      const limit = 5;
-      const address = TEST_DAO_ADDDRESS;
-      const params: IProposalQueryParams = {
-        limit,
-        sortBy: ProposalSortBy.CREATED_AT,
-        direction: SortDirection.ASC,
-        daoAddressOrEns: address,
-      };
-      const proposals = await client.methods.getProposals(params);
-
-      expect(Array.isArray(proposals)).toBe(true);
-      expect(proposals.length > 0 && proposals.length <= limit).toBe(true);
-    });
-    it("Should get a list of proposals from a dao that has no proposals", async () => {
-      const ctx = new Context(contextParams);
-      const ctxPlugin = ContextPlugin.fromContext(ctx);
-      const client = new ClientAddressList(ctxPlugin);
-      const limit = 5;
-      const address = TEST_NON_EXISTING_ADDRESS;
-      const params: IProposalQueryParams = {
-        limit,
-        sortBy: ProposalSortBy.CREATED_AT,
-        direction: SortDirection.ASC,
-        daoAddressOrEns: address,
-      };
-      const proposals = await client.methods.getProposals(params);
-
-      expect(Array.isArray(proposals)).toBe(true);
-      expect(proposals.length === 0).toBe(true);
-    });
-    it("Should get a list of proposals from an invalid address", async () => {
-      const ctx = new Context(contextParams);
-      const ctxPlugin = ContextPlugin.fromContext(ctx);
-      const client = new ClientAddressList(ctxPlugin);
-      const limit = 5;
-      const address = TEST_INVALID_ADDRESS;
-      const params: IProposalQueryParams = {
-        limit,
-        sortBy: ProposalSortBy.CREATED_AT,
-        direction: SortDirection.ASC,
-        daoAddressOrEns: address,
-      };
-      await expect(() => client.methods.getProposals(params)).rejects.toThrow(
-        new InvalidAddressOrEnsError(),
-      );
-    });
-    it("Should get the settings of a plugin given a plugin instance address", async () => {
-      const ctx = new Context(contextParams);
-      const ctxPlugin = ContextPlugin.fromContext(ctx);
-      const client = new ClientAddressList(ctxPlugin);
-
-      const pluginAddress: string = TEST_PLUGIN_ADDRESS;
-      const settings = await client.methods.getSettings(pluginAddress);
-
-      expect(settings === null).toBe(false);
-      if (settings) {
-        expect(typeof settings.minDuration).toBe("number");
-        expect(typeof settings.minSupport).toBe("number");
-        expect(typeof settings.minTurnout).toBe("number");
-      }
     });
   });
 });
