+++ javascript/modules/client/test/integration/constants.ts
@@ -1,0 +1,99 @@
+import { Wallet } from "@ethersproject/wallet";
+import { ContextParams } from "../../src";
+
+const IPFS_API_KEY = process.env.IPFS_API_KEY ||
+  Buffer.from(
+    "YjQ3N1JoRUNmOHM4c2RNN1hya0xCczJ3SGM0a0NNd3BiY0ZDNTVLdCAg==",
+    "base64",
+  ).toString().trim();
+
+export const web3endpoints = {
+  working: [
+    "https://mainnet.infura.io/v3/94d2e8caf1bc4c4884af830d96f927ca",
+    "https://cloudflare-eth.com/",
+  ],
+  failing: ["https://bad-url-gateway.io/"],
+};
+
+const ipfsEndpoints = {
+  working: [
+    {
+      url: "https://testing-ipfs-0.aragon.network/api/v0",
+      headers: {
+        "X-API-KEY": IPFS_API_KEY || "",
+      },
+    },
+    {
+      url: "http://localhost:5001",
+    },
+    {
+      url: "http://localhost:5002",
+    },
+    {
+      url: "http://localhost:5003",
+    },
+  ],
+  failing: [
+    {
+      url: "https://bad-url-gateway.io/",
+    },
+  ],
+};
+
+const grapqhlEndpoints = {
+  working: [
+    {
+      url:
+        "https://api.thegraph.com/subgraphs/name/aragon/aragon-zaragoza-goerli",
+    },
+  ],
+  failing: [{ url: "https://bad-url-gateway.io/" }],
+};
+
+export const TEST_WALLET =
+  "0xdf57089febbacf7ba0bc227dafbffa9fc08a93fdc68e1e42411a14efcf23656e";
+export const TEST_ERC20_PROPOSAL_ID =
+  "0x3e16a08ad61500b2bc77ddcc04387f0fabc04469_0x0";
+export const TEST_ERC20_DAO_ADDRESS =
+  "0x1cab6f621a41438639e1f1b51c274ae65d41b8cb";
+export const TEST_DAO_ADDRESS = "0xa40fd495c454d19c87c4d615100307f0283c017c";
+export const TEST_NO_BALANCES_DAO_ADDRESS =
+  "0xacb269a20d28f4e6b9fc75353cdf24418117859d";
+export const TEST_INVALID_ADDRESS = "0x1nv4l1d_4ddr355";
+export const TEST_NON_EXISTING_ADDRESS =
+  "0x1234567890123456789012345678901234567890";
+export const TEST_ERC20_PLUGIN_ADDRESS =
+  "0x3e16a08ad61500b2bc77ddcc04387f0fabc04469";
+export const TEST_ADDRESSLIST_PROPOSAL_ID =
+  "0x0bcaeb2da9d8ce246ac0864f081b04d4380b6b9c_0x0";
+export const TEST_ADDRESSLIST_DAO_ADDDRESS =
+  "0xa40fd495c454d19c87c4d615100307f0283c017c";
+export const TEST_ADDRESSLIST_PLUGIN_ADDRESS =
+  "0x0bcaeb2da9d8ce246ac0864f081b04d4380b6b9c";
+
+export const contextParams: ContextParams = {
+  network: "mainnet",
+  signer: new Wallet(TEST_WALLET),
+  daoFactoryAddress: "0x0123456789012345678901234567890123456789",
+  web3Providers: web3endpoints.working,
+  ipfsNodes: ipfsEndpoints.working,
+  graphqlNodes: grapqhlEndpoints.working,
+};
+
+export const contextParamsLocalChain: ContextParams = {
+  network: 31337,
+  signer: new Wallet(TEST_WALLET),
+  daoFactoryAddress: "0xf8065dD2dAE72D4A8e74D8BB0c8252F3A9acE7f9",
+  web3Providers: ["http://localhost:8545"],
+  ipfsNodes: ipfsEndpoints.working,
+  graphqlNodes: grapqhlEndpoints.working,
+};
+
+export const contextParamsFailing: ContextParams = {
+  network: "mainnet",
+  signer: new Wallet(TEST_WALLET),
+  daoFactoryAddress: "0x0123456789012345678901234567890123456789",
+  web3Providers: web3endpoints.failing,
+  ipfsNodes: ipfsEndpoints.failing,
+  graphqlNodes: grapqhlEndpoints.failing,
+};
